+++ sertop/sertop_sexp.ml
@@ -32,7 +32,8 @@ type ser_opts =
 ; printer  : Sertop_ser.ser_printer
 
 ; debug    : bool
-; allow_sprop: bool
+; allow_sprop : bool
+; indices_matter : bool
 ; print0   : bool
 ; lheader  : bool                (* Print lenght header (deprecated)           *)
 
@@ -132,6 +133,7 @@ let ser_loop ser_opts =
         ; ml_load      = None
         ; debug        = ser_opts.debug
         ; allow_sprop  = ser_opts.allow_sprop
+        ; indices_matter = ser_opts.indices_matter
         })
   in